+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!-- Please add changes under the Unreleased section that reads 'No current changes' otherwise -->
 
+# Unreleased
+
+- Ensure that committed in-session interface code is being updated
+
 # v4.8.2
 
 - Fix issue where extra data from schema is not reloaded from index